refactor: step and loss classes move to own files
