+++ mava/components/jax/training/__init__.py
@@ -15,7 +15,7 @@
 
 """Trainer components for Mava systems."""
 from mava.components.jax.training.advantage_estimation import GAE
-from mava.components.jax.training.base import Batch, Loss, Step, TrainingState, Utility
+from mava.components.jax.training.base import Batch, TrainingState, Utility
 from mava.components.jax.training.losses import MAPGWithTrustRegionClippingLoss
 from mava.components.jax.training.model_updating import (
     MAPGEpochUpdate,
