+++ mava/components/jax/training/base.py
@@ -51,33 +51,3 @@ class Utility(Component):
     @abc.abstractmethod
     def on_training_utility_fns(self, trainer: SystemTrainer) -> None:
         """[summary]"""
-
-
-class Loss(Component):
-    @abc.abstractmethod
-    def on_training_loss_fns(self, trainer: SystemTrainer) -> None:
-        """[summary]"""
-
-    @staticmethod
-    def name() -> str:
-        """_summary_
-
-        Returns:
-            _description_
-        """
-        return "loss"
-
-
-class Step(Component):
-    @abc.abstractmethod
-    def on_training_step_fn(self, trainer: SystemTrainer) -> None:
-        """[summary]"""
-
-    @staticmethod
-    def name() -> str:
-        """_summary_
-
-        Returns:
-            _description_
-        """
-        return "sgd_step"
